# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A backend service written in Go with Gin framework",
        "title": "LONDONWINNER",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/report/daily-transactions": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves transaction statistics for the current day",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get daily transaction statistics",
                "responses": {}
            }
        },
        "/api/v1/report/draw": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get daily transaction report",
                "responses": {}
            }
        },
        "/api/v1/report/expected-transaction": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Stores expected revenue and transaction count",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Add an expected transaction",
                "parameters": [
                    {
                        "description": "Expected Transaction Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ExpectedTransactionRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/report/general": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Generate a general report for the given date range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Generate general report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "From date (YYYY-MM-DD)",
                        "name": "fromDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "To date (YYYY-MM-DD)",
                        "name": "toDate",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/report/hourly-transaction-report": {
            "get": {
                "description": "Get detailed hourly transaction report for a specific date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get hourly transaction report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date in YYYY-MM-DD format",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/report/monthly": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Generate revenue and transaction statistics grouped by month",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Generate monthly report",
                "responses": {}
            }
        },
...
        "/api/v1/report/product-daily-transaction-report": {
            "get": {
                "description": "Get detailed daily transaction report filtered by product and date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get daily transaction report by product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID (UUID)",
                        "name": "productId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/report/product-hourly-transaction-report": {
            "get": {
                "description": "Get detailed hourly transaction report filtered by product and date range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get hourly transaction report by product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID (UUID)",
                        "name": "productId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "From Date (YYYY-MM-DD HH:mm:ss)",
                        "name": "fromDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "To Date (YYYY-MM-DD HH:mm:ss)",
                        "name": "toDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by clause (e.g., tot_successful_trx_amount DESC)",
                        "name": "orderBy",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/report/product-monthly-transaction-report": {
            "get": {
                "description": "Get detailed monthly transaction report filtered by product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get monthly transaction report by product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID (UUID)",
                        "name": "productId",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/report/top-ten": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves the top 10 transactions based on total transaction amount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get top 10 transactions",
                "responses": {}
            }
        },
        "/api/v1/report/top-ten/recent": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves the top 10 most recent successful transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get top 10 recent successful transactions",
                "responses": {}
            }
        },
        "/api/v1/report/transaction": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Generate a transaction report for the given date range, optionally filtered by product ID and status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Generate transaction report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "From date (YYYY-MM-DD)",
                        "name": "fromDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "To date (YYYY-MM-DD)",
                        "name": "toDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "productId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Transaction status (PENDING, FAILED, SUCCESS)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/report/weekly": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Generate revenue and transaction statistics grouped by week",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Generate weekly report",
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "models.ExpectedTransactionRequest": {
            "type": "object",
            "required": [
                "revenue",
                "transactions"
            ],
            "properties": {
                "revenue": {
                    "type": "number"
                },
                "transactions": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}